transform csvs to json at build time & use new static build api
