+++ web/.env.example
@@ -1,1 +1,2 @@
 ENABLE_MAP=false
+STATES=nevada
