+++ web/.vscode/settings.json
@@ -1,0 +1,3 @@
+{
+  "debug.node.autoAttach": "on"
+}
