+++ web/package.json
@@ -4,18 +4,20 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "next build",
-    "serve": "npx serve out",
+    "build": "next build && next export",
+    "serve": "serve out",
     "start": "next start",
     "type-check": "tsc --pretty --noEmit",
     "format": "prettier --write src/**/*.{js,jsx,ts,tsx}",
     "test": "jest",
     "ci": "yarn test && yarn type-check && yarn lint",
     "lint": "eslint . --ext ts --ext tsx --ext js",
-    "export": "next build && next export"
+    "export": "next export",
+    "export:debug": "NODE_OPTIONS='--inspect' next export"
   },
   "dependencies": {
     "@types/lodash": "^4.14.149",
+    "csv-loader": "^3.0.2",
     "csv-parse": "^4.8.6",
     "d3": "^5.15.0",
     "dotenv": "^8.2.0",
@@ -24,6 +26,7 @@
     "next": "^9.2.2",
     "react": "16.12.0",
     "react-dom": "16.12.0",
+    "serve": "^11.3.0",
     "theme-ui": "^0.3.1",
     "ts-node": "^8.6.2",
     "typescript": "^3.8.2"
