+++ web/pages/states/[name].tsx
@@ -1,17 +1,8 @@
 import React from "react"
 import { NextPage } from "next"
-import fetch from "isomorphic-unfetch"
 import { Layout } from "../../components/Layout"
 import { UserContext } from "../../components/Context"
 import { Main } from "../../components/Main"
-import { massageResult } from "../../util/precinctData"
-
-// get the csv from wherever
-const getCsv = async () => {
-  const result = await fetch("https://nevada-cranks.herokuapp.com/results")
-  const csv = await result.text()
-  return csv
-}
 
 interface Props {
   // data: PrecinctData
@@ -21,6 +12,7 @@ interface Props {
 
 const Nevada: NextPage<Props> = props => {
   const { data, error } = props
+
   if (error) return <p>error: {error}</p>
   const [selectedPrecinct, setSelectedPrecinct] = React.useState(null)
   return (
@@ -34,12 +26,21 @@ const Nevada: NextPage<Props> = props => {
   )
 }
 
-// eslint-disable-next-line @typescript-eslint/camelcase
-export const unstable_getStaticProps = async (...args) => {
-  console.warn("unstable_getStaticProps", { args })
-  const csv = await getCsv()
-  const data = massageResult(csv)
-  return { props: { data: data as any } }
+export const getStaticPaths = () => {
+  return {
+    paths: (process.env.STATES as string)
+      .split(",")
+      .map(state => `/states/${state}`),
+
+    fallback: false,
+  }
+}
+
+export const getStaticProps = async (...args) => {
+  const state = args[0].params.name
+  const data = require(`../../public/data/${state}.json`)
+
+  return { props: { data } }
 }
 
 export default Nevada
